refactor:maptool:Support 56 bit node IDs
We need it because openstreetmap data recently has passed 2^32 node id value.

Also, running maptool on 32 bit systems won't work at least in following cases:

- a node with osm id=>2^32 is a via member of a turn restriction relation;

- input file has unordered node ids, that's the case, for example, for
overpass turbo exported data.

To avoid confusion, I have disabled maptool build for 32bit systems.

+++ navit/maptool/osm_o5m.c
@@ -49,12 +49,12 @@ static unsigned long long
 get_uval(unsigned char **p)
 {
 	unsigned char c;
-	long long ret=0;
+	unsigned long long ret=0;
 	int shift=0;
 
 	for (;;) {
 		c=*((*p)++);
-		ret+=((long long)c & 0x7f) << shift;
+		ret+=((unsigned long long)c & 0x7f) << shift;
 		if (!(c & 0x80)) 
 			return ret;
 		shift+=7;
@@ -64,9 +64,9 @@ get_uval(unsigned char **p)
 static unsigned long long
 get_sval(unsigned char **p)
 {
-	long long ret=get_uval(p);
+	unsigned long long ret=get_uval(p);
 	if (ret & 1) {
-		return -((ret >> 1)+1);
+		return -((long long)(ret >> 1)+1);
 	} else {
 		return ret >> 1;
 	}
